refactor: Make event handlers first-class citizens in our code
